+++ src/types/UserProjectData.ts
@@ -1,3 +1,5 @@
+import { ProjectResource } from "./ProjectResource";
+
 export interface UserProjectData {
     org : string,
     name : string,